dhcpv4: consolidate all IP options into one file.
